Backport trunk commits r5579,5580 into v1.7 bra...
Backport trunk commits r5579,5580 into v1.7 branch:
* lstopo: fix some warnings about double/uint casts (-Wbad-function-cast)
* Don't use "cpuset" as variable name in files built on FreeBSD
  It would conflict with the FreeBSD cpuset() function.
  Harmless but MPICH doesn't like getting such warnings (-Wshadow)

This commit was SVN r5581.

The following SVN revision numbers were found above:
  r5579 --> b2537cc8fa175520454f14064a67b1222cf57c39
  r5580 --> 3012721a61072c685a7cedd961f76d9f461fae4d

+++ utils/lstopo-draw.c
@@ -261,8 +261,8 @@ RECURSE_BEGIN(obj, border) \
     } else { \
       unsigned found = 0; \
       /* Try to find a fitting rectangle */ \
-      for (rows = (unsigned) floor(sqrt(numsubobjs)); \
-	   rows >= (unsigned) ceil(pow(numsubobjs,0.33)) && rows > 1; \
+      for (rows = (unsigned) (float) floor(sqrt(numsubobjs)); \
+	   rows >= (unsigned) (float) ceil(pow(numsubobjs,0.33)) && rows > 1; \
 	   rows--) { \
         columns = numsubobjs / rows; \
         if (columns > 1 && columns * rows == numsubobjs) { \
